pyproject.toml
[project]
name = "eed-webscrapping-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "EED85", email = "eric.brahmann@eed-solutions.de" }
]
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "cryptography>=44.0.2",
    "duckdb>=1.2.0",
    "gitpython>=3.1.44",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "selenium>=4.30.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Set the maximum line length to 79.
line-length = 100

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["E501"]  # Line too long (let black handle this)
